Abstract image printing into a library.
